Refactor HealthThermometer + Add Custom type DateTime + CustomHealthThermometer
